implemented Fact, FunctionApplicability and some facts

+++ Android/ItHappenedAndroid/app/src/main/java/ru/lod_misis/ithappened/Statistics/Facts/Models/FrequentEventsFactModel.java
@@ -5,17 +5,17 @@ package ru.lod_misis.ithappened.Statistics.Facts.Models;
  */
 
 public class FrequentEventsFactModel {
-    int period;
+    Double period;
     String trackingName;
     String trackingId;
 
-    public FrequentEventsFactModel(int period, String trackingName, String trackingId) {
+    public FrequentEventsFactModel(Double period, String trackingName, String trackingId) {
         this.period = period;
         this.trackingName = trackingName;
         this.trackingId = trackingId;
     }
 
-    public int getPeriod() {
+    public Double getPeriod() {
         return period;
     }
